requirements.txt
customtkinter==5.2.0
Pillow==10.0.0
pyautogui==0.9.54
opencv-python==4.8.0.76
numpy==1.24.4
pywinauto==0.6.8
psutil==5.9.5
configparser==6.0.0
diff-match-patch==20230430
pywin32
winocr
googletrans==3.1.0a0